# openapi.yaml
openapi: 3.0.3
info:
  title: Checkin API
  version: 1.0.0
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 200
        registrationType:
          type: string
          minLength: 1
          maxLength: 50
        firstName:
          type: string
          minLength: 1
          maxLength: 100
        middleName:
          type: string
          minLength: 1
          maxLength: 100
        lastName:
          type: string
          minLength: 1
          maxLength: 100
        country:
          type: string
          minLength: 1
          maxLength: 100
        state:
          type: string
          minLength: 1
          maxLength: 100
        mobileNumber:
          type: string
          minLength: 6
          maxLength: 20
        addressLine1:
          type: string
          minLength: 1
          maxLength: 200
        addressLine2:
          type: string
          minLength: 1
          maxLength: 200
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        createdAt:
          type: string
          format: date-time
    AuthResponse:
      type: object
      properties:
        tokenType:
          type: string
        accessToken:
          type: string
        user:
          $ref: "#/components/schemas/UserResponse"
    AppUserDetailsResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        createdAt:
          type: string
          format: date-time
        lastLoginDate:
          type: string
          format: date-time
    UpdateAppUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
    ForgotPasswordResponse:
      type: object
      properties:
        message:
          type: string
        resetToken:
          type: string
    ResetPasswordRequest:
      type: object
      required:
        - token
        - newPassword
      properties:
        token:
          type: string
        newPassword:
          type: string
          minLength: 8
          maxLength: 200
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    UpdateDetailsRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        registrationType:
          type: string
          minLength: 1
          maxLength: 50
        firstName:
          type: string
          minLength: 1
          maxLength: 100
        middleName:
          type: string
          minLength: 1
          maxLength: 100
        lastName:
          type: string
          minLength: 1
          maxLength: 100
        country:
          type: string
          minLength: 1
          maxLength: 100
        state:
          type: string
          minLength: 1
          maxLength: 100
        mobileNumber:
          type: string
          pattern: '^[0-9+\-()\s]+$'
          minLength: 6
          maxLength: 20
        addressLine1:
          type: string
          minLength: 1
          maxLength: 200
        addressLine2:
          type: string
          minLength: 1
          maxLength: 200
    AccountDetailsResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        createdAt:
          type: string
          format: date-time
        lastLoginDate:
          type: string
          format: date-time
        registrationType:
          type: string
        firstName:
          type: string
        middleName:
          type: string
        lastName:
          type: string
        country:
          type: string
        state:
          type: string
        mobileNumber:
          type: string
        addressLine1:
          type: string
        addressLine2:
          type: string
        hasPaid:
          type: boolean
        paidAt:
          type: string
          format: date-time
    EmergencyContactRequest:
      type: object
      required:
        - mobileNumber
        - email
      properties:
        mobileNumber:
          type: string
          pattern: '^[0-9+\-()\s]+$'
          minLength: 6
          maxLength: 20
        email:
          type: string
          format: email
    EmergencyContactResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        mobileNumber:
          type: string
        email:
          type: string
        contactIndex:
          type: integer
paths:
  /api/auth/register:
    post:
      tags:
        - Auth
      summary: Register a new user
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: Login user
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/user/register:
    post:
      tags:
        - User
      summary: Register a new user (via User controller)
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
  /api/user/forgot-password:
    post:
      tags:
        - User
      summary: Request password reset
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "202":
          description: Reset token generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForgotPasswordResponse"
  /api/user/reset-password:
    post:
      tags:
        - User
      summary: Reset password
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /api/user/me:
    get:
      tags:
        - User
      summary: Get current app user details
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUserDetailsResponse"
  /api/user/details:
    put:
      tags:
        - User
      summary: Update app user details
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppUserRequest"
      responses:
        "200":
          description: User details updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUserDetailsResponse"
  /api/login/register:
    post:
      tags:
        - Login
      summary: Register a new user (via Login controller)
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
  /api/login/login:
    post:
      tags:
        - Login
      summary: Login user (via Login controller)
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/login/me:
    get:
      tags:
        - Login
      summary: Get current account details
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Account details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountDetailsResponse"
  /api/login/details:
    put:
      tags:
        - Login
      summary: Update account details
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDetailsRequest"
      responses:
        "200":
          description: Account details updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountDetailsResponse"
  /api/login/forgot-password:
    post:
      tags:
        - Login
      summary: Request password reset (via Login controller)
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "202":
          description: Reset token generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForgotPasswordResponse"
  /api/login/reset-password:
    post:
      tags:
        - Login
      summary: Reset password (via Login controller)
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /api/emergency-contacts:
    get:
      tags:
        - Emergency Contacts
      summary: List emergency contacts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of emergency contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmergencyContactResponse"
    post:
      tags:
        - Emergency Contacts
      summary: Add emergency contact
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmergencyContactRequest"
      responses:
        "201":
          description: Emergency contact created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyContactResponse"
  /api/emergency-contacts/{id}:
    put:
      tags:
        - Emergency Contacts
      summary: Update emergency contact
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmergencyContactRequest"
      responses:
        "200":
          description: Emergency contact updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyContactResponse"
    delete:
      tags:
        - Emergency Contacts
      summary: Delete emergency contact
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Emergency contact deleted