# src/main/java/com/checkin/dto/MessageResponse.java
package com.checkin.dto;

public class MessageResponse {
	private final String message;

	public MessageResponse(String message) {
		this.message = message;
	}

	public String getMessage() {
		return message;
	}
}